backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
flask-limiter==3.7.0
flasgger==0.9.7.1
mysql-connector-python==8.4.0
bcrypt==4.1.3
python-dotenv==1.0.1
werkzeug==3.0.3
pytest==8.2.2
pytest-flask==1.3.0
